+++ benches/normalize.rs
@@ -13,7 +13,7 @@ fn normalize_process() {
         libc::dlopen as Addr,
         libc::fopen as Addr,
         normalize_process as Addr,
-        c_api::blazesym_find_addrs as Addr,
+        c_api::blaze_find_addrs as Addr,
     ];
     let () = addrs.sort();
