+++ src/c_api/mod.rs
@@ -1,7 +1,10 @@
 #[allow(non_camel_case_types)]
+mod inspect;
+#[allow(non_camel_case_types)]
 mod normalize;
 #[allow(non_camel_case_types)]
 mod symbolize;
 
+pub use inspect::*;
 pub use normalize::*;
 pub use symbolize::*;
