+++ src/c_api/symbolize.rs
@@ -18,9 +18,7 @@ use crate::log::warn;
 use crate::util::slice_from_user_array;
 use crate::Addr;
 use crate::BlazeSymbolizer;
-use crate::SymbolInfo;
 use crate::SymbolSrcCfg;
-use crate::SymbolType;
 use crate::SymbolizedResult;
 use crate::SymbolizerFeature;
 
@@ -507,191 +505,6 @@ pub unsafe extern "C" fn blazesym_result_free(results: *const blazesym_result) {
     }
 
     let raw_buf_with_sz = unsafe { (results as *mut u8).offset(-(mem::size_of::<u64>() as isize)) };
-    let sz = unsafe { *(raw_buf_with_sz as *mut u64) } as usize + mem::size_of::<u64>();
-    unsafe { dealloc(raw_buf_with_sz, Layout::from_size_align(sz, 8).unwrap()) };
-}
-
-#[repr(C)]
-#[derive(Debug)]
-pub struct blazesym_sym_info {
-    pub name: *const c_char,
-    pub address: Addr,
-    pub size: usize,
-    pub file_offset: u64,
-    pub obj_file_name: *const c_char,
-    pub sym_type: blazesym_sym_type,
-}
-
-/// Convert [`SymbolInfo`] objects as returned by
-/// [`BlazeSymbolizer::find_addrs`] to a C array.
-fn convert_syms_list_to_c(syms_list: Vec<Vec<SymbolInfo>>) -> *const *const blazesym_sym_info {
-    let mut sym_cnt = 0;
-    let mut str_buf_sz = 0;
-
-    for syms in &syms_list {
-        sym_cnt += syms.len() + 1;
-        for sym in syms {
-            str_buf_sz += sym.name.len() + 1;
-            if let Some(fname) = sym.obj_file_name.as_ref() {
-                str_buf_sz += AsRef::<OsStr>::as_ref(fname).as_bytes().len() + 1;
-            }
-        }
-    }
-
-    let array_sz = ((mem::size_of::<*const u64>() * syms_list.len() + mem::size_of::<u64>() - 1)
-        % mem::size_of::<u64>())
-        * mem::size_of::<u64>();
-    let sym_buf_sz = mem::size_of::<blazesym_sym_info>() * sym_cnt;
-    let buf_size = array_sz + sym_buf_sz + str_buf_sz;
-    let raw_buf_with_sz =
-        unsafe { alloc(Layout::from_size_align(buf_size + mem::size_of::<u64>(), 8).unwrap()) };
-
-    unsafe { *(raw_buf_with_sz as *mut u64) = buf_size as u64 };
-
-    let raw_buf = unsafe { raw_buf_with_sz.add(mem::size_of::<u64>()) };
-    let mut syms_ptr = raw_buf as *mut *mut blazesym_sym_info;
-    let mut sym_ptr = unsafe { raw_buf.add(array_sz) } as *mut blazesym_sym_info;
-    let mut str_ptr = unsafe { raw_buf.add(array_sz + sym_buf_sz) } as *mut c_char;
-
-    for syms in syms_list {
-        unsafe { *syms_ptr = sym_ptr };
-        for SymbolInfo {
-            name,
-            address,
-            size,
-            sym_type,
-            file_offset,
-            obj_file_name,
-        } in syms
-        {
-            let name_ptr = str_ptr.cast();
-            unsafe { ptr::copy_nonoverlapping(name.as_ptr().cast(), str_ptr, name.len()) };
-            str_ptr = unsafe { str_ptr.add(name.len()) };
-            unsafe { *str_ptr = 0 };
-            str_ptr = unsafe { str_ptr.add(1) };
-            let obj_file_name = if let Some(fname) = obj_file_name.as_ref() {
-                let fname = AsRef::<OsStr>::as_ref(fname).as_bytes();
-                let obj_fname_ptr = str_ptr;
-                unsafe { ptr::copy_nonoverlapping(fname.as_ptr().cast(), str_ptr, fname.len()) };
-                str_ptr = unsafe { str_ptr.add(fname.len()) };
-                unsafe { *str_ptr = 0 };
-                str_ptr = unsafe { str_ptr.add(1) };
-                obj_fname_ptr
-            } else {
-                ptr::null()
-            };
-
-            unsafe {
-                (*sym_ptr) = blazesym_sym_info {
-                    name: name_ptr,
-                    address,
-                    size,
-                    sym_type: match sym_type {
-                        SymbolType::Function => blazesym_sym_type::BLAZESYM_SYM_T_FUNC,
-                        SymbolType::Variable => blazesym_sym_type::BLAZESYM_SYM_T_VAR,
-                        SymbolType::Unknown => blazesym_sym_type::BLAZESYM_SYM_T_UNKNOWN,
-                    },
-                    file_offset,
-                    obj_file_name,
-                }
-            };
-            sym_ptr = unsafe { sym_ptr.add(1) };
-        }
-        unsafe {
-            (*sym_ptr) = blazesym_sym_info {
-                name: ptr::null(),
-                address: 0,
-                size: 0,
-                sym_type: blazesym_sym_type::BLAZESYM_SYM_T_UNKNOWN,
-                file_offset: 0,
-                obj_file_name: ptr::null(),
-            }
-        };
-        sym_ptr = unsafe { sym_ptr.add(1) };
-
-        syms_ptr = unsafe { syms_ptr.add(1) };
-    }
-
-    raw_buf as *const *const blazesym_sym_info
-}
-
-
-/// The types of symbols.
-///
-/// This type is used to choice what type of symbols you like to find
-/// and indicate the types of symbols found.
-#[repr(C)]
-#[allow(unused)]
-#[derive(Copy, Clone, Debug)]
-pub enum blazesym_sym_type {
-    /// You want to find a symbol of any type.
-    BLAZESYM_SYM_T_UNKNOWN,
-    /// The returned symbol is a function, or you want to find a function.
-    BLAZESYM_SYM_T_FUNC,
-    /// The returned symbol is a variable, or you want to find a variable.
-    BLAZESYM_SYM_T_VAR,
-}
-
-
-/// Find the addresses of a list of symbols.
-///
-/// Return an array with the same size as the input names. The caller should
-/// free the returned array by calling [`blazesym_syms_list_free()`].
-///
-/// Every name in the input name list may have more than one address.
-/// The respective entry in the returned array is an array containing
-/// all addresses and ended with a null (0x0).
-///
-/// # Safety
-///
-/// The returned pointer should be free by [`blazesym_syms_list_free()`].
-#[no_mangle]
-pub unsafe extern "C" fn blazesym_find_addrs(
-    symbolizer: *mut blazesym,
-    cfg: *const blazesym_sym_src_cfg,
-    names: *const *const c_char,
-    name_cnt: usize,
-) -> *const *const blazesym_sym_info {
-    // SAFETY: The caller ensures that the pointer is valid.
-    let symbolizer = unsafe { &*symbolizer };
-    // SAFETY: The caller ensures that the pointer is valid.
-    let cfg = SymbolSrcCfg::from(unsafe { &*cfg });
-    // SAFETY: The caller ensures that the pointer is valid and the count
-    //         matches.
-    let names = unsafe { slice_from_user_array(names, name_cnt) };
-    let names = names
-        .iter()
-        .map(|&p| {
-            // SAFETY: The caller ensures that the pointer is valid.
-            unsafe { CStr::from_ptr(p) }.to_str().unwrap()
-        })
-        .collect::<Vec<_>>();
-    let result = symbolizer.find_addrs(&cfg, &names);
-    match result {
-        Ok(syms) => convert_syms_list_to_c(syms),
-        Err(_err) => {
-            // TODO: Errors should probably be surfaced.
-            #[cfg(debug_assertions)]
-            eprintln!("failed to find {name_cnt} symbols: {_err}");
-            ptr::null()
-        }
-    }
-}
-
-/// Free an array returned by [`blazesym_find_addrs`].
-///
-/// # Safety
-///
-/// The pointer must be returned by [`blazesym_find_addrs`].
-///
-#[no_mangle]
-pub unsafe extern "C" fn blazesym_syms_list_free(syms_list: *const *const blazesym_sym_info) {
-    if syms_list.is_null() {
-        return
-    }
-
-    let raw_buf_with_sz =
-        unsafe { (syms_list as *mut u8).offset(-(mem::size_of::<u64>() as isize)) };
     let sz = unsafe { *(raw_buf_with_sz as *mut u64) } as usize + mem::size_of::<u64>();
     unsafe { dealloc(raw_buf_with_sz, Layout::from_size_align(sz, 8).unwrap()) };
 }
